[ExcelAnalyzer] update autoloading
Fixes issues in development mode where we were seeing:
> superclass mismatch for class ExcelAnalyzerNotifier

Fixes #1869
