+++ lib/excel_analyzer.rb
@@ -1,3 +1,11 @@
+Rails.application.config.before_initialize do
+  loader = Rails.autoloaders.main
+
+  Dir[File.join(File.dirname(__FILE__), 'excel_analyzer', '**/')].each do
+    loader.push_dir(_1, namespace: ExcelAnalyzer)
+  end
+end
+
 ExcelAnalyzer.on_hidden_metadata = ->(attachment_blob, metadata) do
   foi_attachment = FoiAttachment.joins(:file_blob).
     find_by(active_storage_blobs: { id: attachment_blob })
@@ -16,32 +24,5 @@ ExcelAnalyzer.on_hidden_metadata = ->(attachment_blob, metadata) do
     }
   )
 
-  ExcelAnalyzerNotifier.report(foi_attachment, metadata).deliver_now
-end
-
-Rails.configuration.to_prepare do
-  class ExcelAnalyzerNotifier < ApplicationMailer
-    include Rails.application.routes.url_helpers
-    default_url_options[:host] = AlaveteliConfiguration.domain
-
-    def report(foi_attachment, metadata)
-      @foi_attachment = foi_attachment
-      @incoming_message = foi_attachment.incoming_message
-      @metadata = metadata
-
-      from = email_address_with_name(
-        blackhole_email, 'WhatDoTheyKnow.com Excel Analyzer report'
-      )
-
-      headers['X-WDTK-Contact'] = 'wdtk-excel-analyzer-report'
-      headers['X-WDTK-CaseRef'] = @foi_attachment.id
-
-      mail(
-        from: from,
-        to: ENV['EXCEL_ANALYZER_NOTIFICATION_EMAIL'],
-        subject: _('ExcelAnalyzer: hidden data detected [{{reference}}]',
-                   reference: @foi_attachment.id)
-      )
-    end
-  end
+  ExcelAnalyzer::NotifierMailer.report(foi_attachment, metadata).deliver_now
 end
